fix tests

+++ torchgeo/trainers/detection.py
@@ -272,7 +272,7 @@ class ObjectDetectionTask(BaseTask):
         batch_size = x.shape[0]
         assert 'bbox_xyxy' in batch, 'bbox_xyxy is required for object detection.'
         y = [
-            {'boxes': batch['bbox_xyxy'][i], 'labels': batch['label'][i]}
+            {'boxes': batch['bbox_xyxy'][i], 'labels': batch['label'][i].to(torch.long)}
             for i in range(batch_size)
         ]
         y_hat = self(x)
@@ -333,7 +333,7 @@ class ObjectDetectionTask(BaseTask):
         batch_size = x.shape[0]
         assert 'bbox_xyxy' in batch, 'bbox_xyxy is required for object detection.'
         y = [
-            {'boxes': batch['bbox_xyxy'][i], 'labels': batch['label'][i]}
+            {'boxes': batch['bbox_xyxy'][i], 'labels': batch['label'][i].to(torch.long)}
             for i in range(batch_size)
         ]
         y_hat = self(x)
