+++ torchgeo/trainers/instance_segmentation.py
@@ -166,7 +166,7 @@ class InstanceSegmentationTask(BaseTask):
         x = batch['image']
         y = {
             'boxes': batch['bbox_xyxy'],
-            'labels': batch['label'],
+            'labels': batch['label'].to(torch.long),
             'masks': batch['mask'],
         }
         y_hat = self(x.unbind())
@@ -228,7 +228,7 @@ class InstanceSegmentationTask(BaseTask):
         x = batch['image']
         y = {
             'boxes': batch['bbox_xyxy'],
-            'labels': batch['label'],
+            'labels': batch['label'].to(torch.long),
             'masks': batch['mask'],
         }
         y_hat = self(x.unbind())
